Part 2
I'm not sure if I need to commit this, but I am just in case. I have
implemented the changes to add and remove, which now adjust the list
size when necessary. I also created a new method called void listCopy,
which is used in the new updated add and remove methods.